backend/requirements-backend.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements-backend.txt --group backend
aerich==0.8.2
    # via trustenroll (pyproject.toml:backend)
aiofiles==24.1.0
    # via
    #   trustenroll (pyproject.toml:backend)
    #   fastapi-admin
aioredis==2.0.1
    # via
    #   trustenroll (pyproject.toml:backend)
    #   fastapi-admin
aiosqlite==0.21.0
    # via
    #   trustenroll (pyproject.toml:backend)
    #   tortoise-orm
annotated-types==0.7.0
    # via
    #   trustenroll (pyproject.toml:backend)
    #   pydantic
anyio==4.9.0
    # via
    #   trustenroll (pyproject.toml:backend)
    #   asyncclick
    #   starlette
    #   watchfiles
async-timeout==5.0.1
    # via
    #   trustenroll (pyproject.toml:backend)
    #   aioredis
    #   asyncpg
    #   redis
asyncclick==8.1.8
    # via
    #   trustenroll (pyproject.toml:backend)
    #   aerich
asyncpg==0.30.0
    # via trustenroll (pyproject.toml:backend)
babel==2.17.0
    # via
    #   trustenroll (pyproject.toml:backend)
    #   fastapi-admin
bcrypt==4.3.0
    # via
    #   trustenroll (pyproject.toml:backend)
    #   fastapi-admin
click==8.1.8
    # via
    #   trustenroll (pyproject.toml:backend)
    #   uvicorn
colorama==0.4.6
    # via trustenroll (pyproject.toml:backend)
dictdiffer==0.9.0
    # via
    #   trustenroll (pyproject.toml:backend)
    #   aerich
exceptiongroup==1.2.2
    # via anyio
fastapi==0.115.11
    # via
    #   trustenroll (pyproject.toml:backend)
    #   fastapi-admin
    #   fastapi-cache2
fastapi-admin==1.0.4
    # via trustenroll (pyproject.toml:backend)
fastapi-cache2==0.2.2
    # via trustenroll (pyproject.toml:backend)
h11==0.14.0
    # via
    #   trustenroll (pyproject.toml:backend)
    #   uvicorn
httptools==0.6.4
    # via
    #   trustenroll (pyproject.toml:backend)
    #   uvicorn
idna==3.10
    # via
    #   trustenroll (pyproject.toml:backend)
    #   anyio
iso8601==2.1.0
    # via
    #   trustenroll (pyproject.toml:backend)
    #   tortoise-orm
jinja2==3.1.6
    # via
    #   trustenroll (pyproject.toml:backend)
    #   fastapi-admin
markupsafe==3.0.2
    # via
    #   trustenroll (pyproject.toml:backend)
    #   jinja2
passlib==1.7.4
    # via trustenroll (pyproject.toml:backend)
pendulum==3.0.0
    # via
    #   trustenroll (pyproject.toml:backend)
    #   fastapi-admin
    #   fastapi-cache2
pydantic==2.10.6
    # via
    #   trustenroll (pyproject.toml:backend)
    #   aerich
    #   fastapi
pydantic-core==2.27.2
    # via
    #   trustenroll (pyproject.toml:backend)
    #   pydantic
pypika-tortoise==0.5.0
    # via
    #   trustenroll (pyproject.toml:backend)
    #   tortoise-orm
python-dateutil==2.9.0.post0
    # via
    #   trustenroll (pyproject.toml:backend)
    #   pendulum
    #   time-machine
python-dotenv==1.0.1
    # via
    #   trustenroll (pyproject.toml:backend)
    #   uvicorn
python-multipart==0.0.20
    # via
    #   trustenroll (pyproject.toml:backend)
    #   fastapi-admin
pytz==2025.1
    # via
    #   trustenroll (pyproject.toml:backend)
    #   tortoise-orm
pyyaml==6.0.2
    # via
    #   trustenroll (pyproject.toml:backend)
    #   uvicorn
redis==5.2.1
    # via trustenroll (pyproject.toml:backend)
six==1.17.0
    # via
    #   trustenroll (pyproject.toml:backend)
    #   python-dateutil
sniffio==1.3.1
    # via
    #   trustenroll (pyproject.toml:backend)
    #   anyio
starlette==0.46.1
    # via
    #   trustenroll (pyproject.toml:backend)
    #   fastapi
time-machine==2.16.0
    # via
    #   trustenroll (pyproject.toml:backend)
    #   pendulum
tomlkit==0.13.2
    # via trustenroll (pyproject.toml:backend)
tortoise-orm==0.24.2
    # via
    #   trustenroll (pyproject.toml:backend)
    #   aerich
    #   fastapi-admin
typing-extensions==4.12.2
    # via
    #   trustenroll (pyproject.toml:backend)
    #   aioredis
    #   aiosqlite
    #   anyio
    #   fastapi
    #   fastapi-cache2
    #   pydantic
    #   pydantic-core
    #   uvicorn
tzdata==2025.1
    # via
    #   trustenroll (pyproject.toml:backend)
    #   pendulum
uvicorn==0.34.0
    # via
    #   trustenroll (pyproject.toml:backend)
    #   fastapi-admin
    #   fastapi-cache2
uvloop==0.21.0
    # via uvicorn
watchfiles==1.0.4
    # via
    #   trustenroll (pyproject.toml:backend)
    #   uvicorn
websockets==15.0.1
    # via
    #   trustenroll (pyproject.toml:backend)
    #   uvicorn
ua-parser==1.0.1
    # via user-agents
ua-parser-builtins==0.18.0.post1
    # via ua-parser
user-agents==2.2.0
    # via trustenroll (pyproject.toml:backend)
